requirements.txt
discord.py==1.7.3
flask
bs4